sending

+++ src/buffer.cpp
@@ -8,7 +8,7 @@
 
 #include "buffer.h"
 
-Buffer::Buffer(uint8_t *buf, uint64_t buf_size, uint32_t no_bufs) {
+Buffer::Buffer(uint8_t *buf, uint32_t buf_size, uint32_t no_bufs) {
   this->buf = buf;
   this->buf_size = buf_size;
   this->no_bufs = no_bufs;
